Add common interface module.
